option to add in additional curl opts
Reluctantly added this in to allow for adding in curl options that may not be covering by the rest of the library.  Because one of the main goals of this library is to get away from the cURL opt mess and have a more readable library, its only intended to be used as a last resort.