adds support for custom serialization callbacks
Not happy with the built in serialization?  Does the built in serialization not support a mime type that you need?  Now you can just add it in via registerPayloadSerializer.

This method allows you to specify a mime type and a callback/Closure that you wish to use to handle serializing a particular mime type.  Also lets you define a catch all mime type.

Unit tests added