Compose request raw_header.
Mainly done to be able to test against the composed header without a bajor refactor.
This is now similar to the response and although there are other ways to get this information from cURL this now makes HttpFul complete being able to provide both request and response headers.

Note the HTTP version is hard coded this might require additional work but should be efficient for the majority of use cases.