Fix for frameworks that use object proxies
I'm using the Typo3 Flow framework, that creates proxies for all classes.

The proxy is created by extending the Request-class and overrides the constructor. It initializes the original Request-class by calling parent::__construct, but the private __construct method is not in scope. Instead the constructor should be marked as protected.